refactor(agents): extract pre-execution failure handling from cook.rs
Second cook.rs restabilization carve, rebased onto main after the adoption carve
(#9266) merged. cook.rs regrew to 3665 LOC as retry/recovery fixes accreted onto
the run_cook spine. This extracts the attempt-setup and terminal-classification
helpers around the loop into cook_pre_execution.rs (~174 LOC):
materialize_initial_cook_attempt (durable first attempt submission), the
pre-execution failure boundary (retryable_pre_execution_failure,
with_pre_execution_phase, pre_execution_failure_phase/details/report,
record_pre_execution_failure, PreExecutionFailureDetails) that keeps a
lifecycle-owned failure distinct from a provider result, and the terminal-executor
identity helpers (terminal_executor_matches, provider_rotation_attempts,
terminal_executor_identity, TerminalExecutorIdentity) used to decide follow-up rotation.

Behavior-preserving pure move following the agent_task_service module pattern
(mod X; pub use X::*). New module has explicit imports and references cook.rs-owned
types via super::cook::. Widened the spine-called fns + PreExecutionFailureDetails/
TerminalExecutorIdentity structs to pub(crate). Repointed cook_budget's
provider_rotation_attempts import to the new module. Trimmed cook.rs's now-unused
AgentTaskExecutor import.

cook.rs: 3030 -> 2855 LOC (3665 -> 2855 across both cook restabilization carves).
