+++ crates/homeboy-agents/src/agent_task_service/cook.rs
@@ -6,7 +6,6 @@ use serde_json::Value;
 use std::path::PathBuf;
 use std::sync::{mpsc, Arc, Mutex};
 
-use crate::agent_task::AgentTaskExecutor;
 use crate::agent_task_cook_loop::{
     evaluate_cook_loop, AgentTaskCookLoopOptions, AgentTaskCookLoopReport, AgentTaskCookLoopStatus,
 };
@@ -27,6 +26,11 @@ use super::cook_baseline::{
 };
 use super::cook_budget::{
     budget_remaining, execution_budget_usage, reserve_remediation_budget, ExecutionBudgetUsage,
+};
+use super::cook_pre_execution::{
+    materialize_initial_cook_attempt, pre_execution_failure_details, pre_execution_failure_phase,
+    pre_execution_failure_report, record_pre_execution_failure, retryable_pre_execution_failure,
+    terminal_executor_matches, with_pre_execution_phase,
 };
 use super::cook_promotion::{
     attempt_needs_execution, cook_report, finalize_or_load_cook_pr, promote_or_load_attempt,
@@ -799,181 +803,6 @@ where
         Some("cook attempt budget exhausted".to_string()),
         1,
     ))
-}
-
-/// Persist the controller-owned initial attempt before transport preparation so
-/// runner eligibility failures remain addressable through the cook alias.
-fn materialize_initial_cook_attempt(options: &AgentTaskCookServiceOptions) -> Result<()> {
-    if agent_task_lifecycle::run_record_exists(&options.initial_run_id)? {
-        return Ok(());
-    }
-    match agent_task_lifecycle::submit_plan(&options.initial_plan, Some(&options.initial_run_id)) {
-        Ok(_) => {
-            agent_task_lifecycle::record_cook_attempt(&options.cook_id, 1, &options.initial_run_id)
-                .map(|_| ())
-        }
-        Err(error) => {
-            // `submit_plan` persists admission failures before returning them.
-            if agent_task_lifecycle::run_record_exists(&options.initial_run_id)? {
-                agent_task_lifecycle::record_cook_attempt(
-                    &options.cook_id,
-                    1,
-                    &options.initial_run_id,
-                )?;
-            }
-            Err(error)
-        }
-    }
-}
-
-fn retryable_pre_execution_failure(record: &agent_task_lifecycle::AgentTaskRunRecord) -> bool {
-    record.metadata["pre_execution_failure"]["retryable"] == Value::Bool(true)
-}
-
-#[derive(Debug)]
-struct PreExecutionFailureDetails {
-    retryable: bool,
-    phase: Option<String>,
-    classification: Option<String>,
-}
-
-fn with_pre_execution_phase(mut error: Error, phase: &str) -> Error {
-    if !error.details.is_object() {
-        error.details = serde_json::json!({});
-    }
-    error.details["pre_execution_phase"] = Value::String(phase.to_string());
-    error
-}
-
-fn pre_execution_failure_phase<'a>(
-    error: &'a Error,
-    dispatcher: Option<&dyn AgentTaskCookAttemptDispatcher>,
-) -> &'a str {
-    error
-        .details
-        .get("pre_execution_phase")
-        .and_then(Value::as_str)
-        .unwrap_or_else(|| {
-            dispatcher
-                .map(|dispatcher| dispatcher.pre_execution_failure_phase())
-                .unwrap_or("cook_pre_execution")
-        })
-}
-
-fn pre_execution_failure_details(
-    record: Option<&agent_task_lifecycle::AgentTaskRunRecord>,
-    error: &Error,
-) -> PreExecutionFailureDetails {
-    let failure = record.and_then(|record| record.metadata.get("pre_execution_failure"));
-    PreExecutionFailureDetails {
-        retryable: failure
-            .and_then(|failure| failure.get("retryable"))
-            .and_then(Value::as_bool)
-            .unwrap_or(error.retryable == Some(true)),
-        phase: failure
-            .and_then(|failure| failure.get("phase"))
-            .and_then(Value::as_str)
-            .map(str::to_string),
-        classification: failure
-            .and_then(|failure| failure.get("failure_classification"))
-            .and_then(Value::as_str)
-            .map(str::to_string),
-    }
-}
-
-fn pre_execution_failure_report(
-    cook_id: String,
-    attempts: Vec<AgentTaskCookAttemptReport>,
-    failure: PreExecutionFailureDetails,
-    error: Error,
-) -> AgentTaskRunResult<AgentTaskCookReport> {
-    let phase = failure.phase.as_deref().unwrap_or("cook_pre_execution");
-    let classification = failure.classification.as_deref().unwrap_or("unknown");
-    let mut report = cook_report(
-        cook_id,
-        "pre_execution_failure",
-        attempts,
-        None,
-        Some(format!(
-            "pre-provider failure in phase `{phase}` classified as `{classification}`: {error}"
-        )),
-        1,
-    );
-    report.value.terminal_phase = failure.phase;
-    report.value.terminal_failure_classification = failure.classification;
-    report
-}
-
-/// Pre-execution failures happen before a provider can receive work. Persist a
-/// normal terminal run so the Cook alias can expose its complete retry history.
-fn record_pre_execution_failure(
-    plan: &AgentTaskPlan,
-    run_id: &str,
-    error: &Error,
-    phase: &str,
-) -> Result<()> {
-    if !agent_task_lifecycle::run_record_exists(run_id)? {
-        agent_task_lifecycle::submit_plan(plan, Some(run_id))?;
-    }
-    agent_task_lifecycle::record_pre_execution_failure(run_id, plan, phase, error)?;
-    Ok(())
-}
-
-fn terminal_executor_matches(
-    aggregate: &crate::agent_task_scheduler::AgentTaskAggregate,
-    follow_up: &AgentTaskExecutor,
-) -> Option<bool> {
-    let outcome = aggregate.outcomes.last()?;
-    let terminal = terminal_executor_identity(outcome)?;
-    Some(
-        terminal.backend == follow_up.backend
-            && terminal.selector == follow_up.selector
-            && terminal.model.as_deref() == follow_up.model(),
-    )
-}
-
-pub(crate) fn provider_rotation_attempts(
-    outcome: &crate::agent_task::AgentTaskOutcome,
-) -> Option<Vec<crate::agent_task_scheduler::AgentTaskProviderRotationAttempt>> {
-    serde_json::from_value(
-        outcome
-            .metadata
-            .pointer("/provider_rotation/attempts")?
-            .clone(),
-    )
-    .ok()
-}
-
-struct TerminalExecutorIdentity {
-    backend: String,
-    selector: Option<String>,
-    model: Option<String>,
-}
-
-fn terminal_executor_identity(
-    outcome: &crate::agent_task::AgentTaskOutcome,
-) -> Option<TerminalExecutorIdentity> {
-    if let Some(attempt) =
-        provider_rotation_attempts(outcome).and_then(|attempts| attempts.last().cloned())
-    {
-        return Some(TerminalExecutorIdentity {
-            backend: attempt.backend,
-            selector: attempt.selector,
-            model: attempt.model,
-        });
-    }
-    let executor = outcome.metadata.get("executor")?;
-    Some(TerminalExecutorIdentity {
-        backend: executor.get("backend")?.as_str()?.to_string(),
-        selector: executor
-            .get("selector")
-            .and_then(Value::as_str)
-            .map(str::to_string),
-        model: executor
-            .get("model")
-            .and_then(Value::as_str)
-            .map(str::to_string),
-    })
 }
 
 #[cfg(test)]
