+++ crates/homeboy-agents/src/agent_task_service/mod.rs
@@ -9,6 +9,7 @@ mod cook;
 mod cook_adoption;
 mod cook_baseline;
 mod cook_budget;
+mod cook_pre_execution;
 mod cook_promotion;
 mod cook_recipe;
 mod discovery;
@@ -20,6 +21,7 @@ pub use cook::*;
 pub use cook_adoption::*;
 pub use cook_baseline::*;
 pub use cook_budget::*;
+pub use cook_pre_execution::*;
 pub use cook_promotion::*;
 pub use cook_recipe::*;
 pub use discovery::*;
